Create constant for positive int type

+++ src/compile/txt_literal.rs
@@ -7,12 +7,12 @@ use crate::compile::letter::Letter::*;
 use crate::compile::var_uint::DecodeError::TooLarge;
 use crate::op::Op;
 use crate::NR;
+use crate::UINT;
 
-#[inline]
-fn encode_uint_vec(nrs: &[u64]) -> Vec<Letter> {
+fn encode_uint_vec(nrs: &[UINT]) -> Vec<Letter> {
     unimplemented!() //TODO @mark: TEMPORARY! REMOVE THIS!
 }
 
-pub fn decode_uint_vec(letters: &[Letter]) -> Vec<u64> {
+pub fn decode_uint_vec(letters: &[Letter]) -> Vec<UINT> {
     unimplemented!()
 }
